fjerner deprecated kode tilknyttet gammelt design for skolepenger på vedtak og beregningsiden
