+++ src/frontend/Komponenter/Behandling/VedtakOgBeregning/Skolepenger/Felles/vedtaksvalidering.ts
@@ -4,14 +4,15 @@ import {
     ISkoleårsperiodeSkolepenger,
     SkolepengerUtgift,
 } from '../../../../../App/typer/vedtak';
-import { InnvilgeVedtakForm } from '../Felles/typer';
+import { InnvilgeVedtakForm } from '../InnvilgeVedtak/InnvilgeVedtak';
 import {
     erMånedÅrEtterEllerLik,
     Intervall,
     månedÅrTilDate,
     overlapper,
 } from '../../../../../App/utils/dato';
-import { beregnSkoleår } from '../Felles/skoleår';
+import { beregnSkoleår, validerSkoleår } from './skoleår';
+import { validerGyldigTallverdi } from '../../Felles/utils';
 
 const periodeSkolepengerFeil: FormErrors<IPeriodeSkolepenger> = {
     studietype: undefined,
@@ -27,62 +28,82 @@ const periodeUtgiftFeil: FormErrors<SkolepengerUtgift> = {
     stønad: undefined,
 };
 
-export const validerInnvilgetVedtakForm = ({
+export const validerSkoleårsperioderMedBegrunnelse = ({
     skoleårsperioder,
     begrunnelse,
 }: InnvilgeVedtakForm): FormErrors<InnvilgeVedtakForm> => {
+    const validerteSkoleår = validerSkoleår(skoleårsperioder);
+
+    //TODO: Sett denne feilmeldingen på det skoleåret som er årsaken til duplikatfeil, hvis ønskelig
+    if (!validerteSkoleår.gyldig) {
+        return { skoleårsperioder: [], begrunnelse: validerteSkoleår.årsak };
+    }
+
     return {
-        skoleårsperioder: validerSkoleårsperioderSkolepenger(skoleårsperioder, true),
+        skoleårsperioder: validerSkoleårsperioder(skoleårsperioder),
         begrunnelse: !harVerdi(begrunnelse) ? 'Mangelfull utfylling av begrunnelse' : undefined,
     };
 };
 
-export const validerInnvilgetVedtakFormUtenUtgifter = ({
+export const validerSkoleårsperioderUtenBegrunnelse = ({
     skoleårsperioder,
-    begrunnelse,
 }: InnvilgeVedtakForm): FormErrors<InnvilgeVedtakForm> => {
-    return {
-        skoleårsperioder: validerSkoleårsperioderSkolepenger(skoleårsperioder, false),
-        begrunnelse: !harVerdi(begrunnelse) ? 'Mangelfull utfylling av begrunnelse' : undefined,
-    };
-};
+    const validerteSkoleår = validerSkoleår(skoleårsperioder);
 
-export const validerSkoleårsperioder = ({
-    skoleårsperioder,
-}: InnvilgeVedtakForm): FormErrors<InnvilgeVedtakForm> => {
+    //TODO: Sett denne feilmeldingen på det skoleåret som er årsaken til duplikatfeil, hvis ønskelig
+    if (!validerteSkoleår.gyldig) {
+        return { skoleårsperioder: [], begrunnelse: validerteSkoleår.årsak };
+    }
+
     return {
-        skoleårsperioder: validerSkoleårsperioderSkolepenger(skoleårsperioder, true),
+        skoleårsperioder: validerSkoleårsperioder(skoleårsperioder),
         begrunnelse: undefined,
     };
 };
 
-export const validerSkoleårsperioderUtenUtgifter = ({
+export const validerSkoleårsperioderUtenBegrunnelseOgUtgiftsperioder = ({
     skoleårsperioder,
 }: InnvilgeVedtakForm): FormErrors<InnvilgeVedtakForm> => {
+    const validerteSkoleår = validerSkoleår(skoleårsperioder);
+
+    //TODO: Sett denne feilmeldingen på det skoleåret som er årsaken til duplikatfeil, hvis ønskelig
+    if (!validerteSkoleår.gyldig) {
+        return { skoleårsperioder: [], begrunnelse: validerteSkoleår.årsak };
+    }
+
     return {
-        skoleårsperioder: validerSkoleårsperioderSkolepenger(skoleårsperioder, false),
+        skoleårsperioder: validerSkoleårsperioderUtenUtgiftsperioder(skoleårsperioder),
         begrunnelse: undefined,
     };
 };
 
-const validerSkoleårsperioderSkolepenger = (
-    perioder: ISkoleårsperiodeSkolepenger[],
-    skalValidereMangelfullUtfyllingAvUtgifter: boolean | undefined
+const validerSkoleårsperioder = (
+    perioder: ISkoleårsperiodeSkolepenger[]
+): FormErrors<ISkoleårsperiodeSkolepenger[]> => {
+    return perioder.map((periode) => {
+        const skoleårsperiodeFeil: FormErrors<ISkoleårsperiodeSkolepenger> = {
+            perioder: validerDelårsperioder(periode.perioder),
+            utgiftsperioder: validerUtgifsperioder(periode.utgiftsperioder),
+            erHentetFraBackend: undefined,
+        };
+        return skoleårsperiodeFeil;
+    });
+};
+
+export const validerSkoleårsperioderUtenUtgiftsperioder = (
+    perioder: ISkoleårsperiodeSkolepenger[]
 ): FormErrors<ISkoleårsperiodeSkolepenger[]> => {
     return perioder.map((periode) => {
         const utgiftsperiodeFeil: FormErrors<ISkoleårsperiodeSkolepenger> = {
-            perioder: validerDelperiodeSkoleår(periode.perioder),
-            utgiftsperioder: validerUtgifter(
-                periode.utgiftsperioder,
-                skalValidereMangelfullUtfyllingAvUtgifter
-            ),
+            perioder: validerDelårsperioder(periode.perioder),
+            utgiftsperioder: [],
             erHentetFraBackend: undefined,
         };
         return utgiftsperiodeFeil;
     });
 };
 
-const validerDelperiodeSkoleår = (
+const validerDelårsperioder = (
     perioder: IPeriodeSkolepenger[]
 ): FormErrors<IPeriodeSkolepenger[]> => {
     let skoleår: number | undefined = undefined;
@@ -92,10 +113,16 @@ const validerDelperiodeSkoleår = (
         if (!studietype) {
             return { ...periodeSkolepengerFeil, studietype: 'Mangelfull utfylling av studietype' };
         }
+        if (!studiebelastning) {
+            return {
+                ...periodeSkolepengerFeil,
+                studiebelastning: 'Mangelfull utfylling av studiebelastning',
+            };
+        }
         if (!årMånedFra || !årMånedTil) {
             return {
                 ...periodeSkolepengerFeil,
-                årMånedFra: 'Mangelfull utfylling av utgiftsperiode',
+                årMånedFra: 'Mangelfull utfylling av gjeldende periode',
             };
         }
         if (!erMånedÅrEtterEllerLik(årMånedFra, årMånedTil)) {
@@ -131,12 +158,6 @@ const validerDelperiodeSkoleår = (
                 };
             }
         }
-        if (!studiebelastning) {
-            return {
-                ...periodeSkolepengerFeil,
-                studiebelastning: 'Mangelfull utfylling av studiebelastning',
-            };
-        }
         if (studiebelastning < 50 || studiebelastning > 100) {
             return {
                 ...periodeSkolepengerFeil,
@@ -147,12 +168,9 @@ const validerDelperiodeSkoleår = (
     });
 };
 
-const validerUtgifter = (
-    perioder: SkolepengerUtgift[],
-    skalValidereMangelfullUtfyllingAvUtgifter: boolean | undefined
-): FormErrors<SkolepengerUtgift[]> => {
+const validerUtgifsperioder = (perioder: SkolepengerUtgift[]): FormErrors<SkolepengerUtgift[]> => {
     return perioder.map((periode) => {
-        const { årMånedFra, utgifter, stønad } = periode;
+        const { årMånedFra, stønad } = periode;
 
         if (!årMånedFra) {
             return {
@@ -164,27 +182,13 @@ const validerUtgifter = (
         if (stønad === undefined || stønad === null) {
             return {
                 ...periodeUtgiftFeil,
-                stønad: 'Mangelfull utfylling av stønad',
+                stønad: 'Mangelfull utfylling av beløp',
             };
         }
-
-        if (skalValidereMangelfullUtfyllingAvUtgifter) {
-            if (!utgifter || utgifter < 1) {
-                return {
-                    ...periodeUtgiftFeil,
-                    utgifter: 'Mangelfull utfylling av utgifter',
-                };
-            }
-
-            if (stønad > utgifter) {
-                return {
-                    ...periodeUtgiftFeil,
-                    stønad: 'Stønad kan ikke være høyere enn utgifter',
-                };
-            }
-        }
-
-        return periodeUtgiftFeil;
+        return {
+            ...periodeUtgiftFeil,
+            stønad: validerGyldigTallverdi(stønad),
+        };
     });
 };
