+++ src/frontend/Komponenter/Behandling/VedtakOgBeregning/Skolepenger/OpphøreVedtak/OpphøreVedtak.tsx
@@ -1,195 +1,145 @@
-import { ISkoleårsperiodeSkolepenger } from '../../../../../App/typer/vedtak';
-import React, { Dispatch, SetStateAction, useMemo } from 'react';
-import styled from 'styled-components';
-import { useBehandling } from '../../../../../App/context/BehandlingContext';
-import { ListState } from '../../../../../App/hooks/felles/useListState';
-import { useApp } from '../../../../../App/context/AppContext';
-import { VEDTAK_OG_BEREGNING } from '../../Felles/konstanter';
-import { beregnSkoleår, GyldigBeregnetSkoleår } from '../Felles/skoleår';
 import {
-    locateIndexToRestorePreviousItemInCurrentItems,
-    oppdaterValideringsfeil,
-} from '../Felles/utils';
-import Delårsperioder from './Delårsperioder';
-import { FormErrors } from '../../../../../App/hooks/felles/useFormState';
-import { InnvilgeVedtakForm } from '../Felles/typer';
-import FjernKnapp from '../../../../../Felles/Knapper/FjernKnapp';
-import TilbakestillKnapp from '../../../../../Felles/Knapper/TilbakestillKnapp';
-import { AGray50 } from '@navikt/ds-tokens/dist/tokens';
-import Utgiftsperioder from './Utgiftsperioder';
+    EBehandlingResultat,
+    IBeregningSkolepengerResponse,
+    ISkoleårsperiodeSkolepenger,
+    IVedtakForSkolepenger,
+    IvedtakForSkolepenger,
+    IVedtakType,
+} from '../../../../../App/typer/vedtak';
+import { Behandling } from '../../../../../App/typer/fagsak';
+import React, { useEffect, useState } from 'react';
+import useFormState, { FormState } from '../../../../../App/hooks/felles/useFormState';
+import { ListState } from '../../../../../App/hooks/felles/useListState';
+import { useBehandling } from '../../../../../App/context/BehandlingContext';
+import styled from 'styled-components';
+import { FieldState } from '../../../../../App/hooks/felles/useFieldState';
+import { useApp } from '../../../../../App/context/AppContext';
+import { byggTomRessurs, Ressurs, RessursStatus } from '../../../../../App/typer/ressurs';
+import { UtregningstabellSkolepenger } from '../Felles/UtregnignstabellSkolepenger';
+import { InnvilgeVedtakForm, tomSkoleårsperiodeSkolepenger } from '../Felles/typer';
+import OpphøreSkolepenger from './OpphøreSkolepenger';
+import { useRedirectEtterLagring } from '../../../../../App/hooks/felles/useRedirectEtterLagring';
+import { v4 as uuidv4 } from 'uuid';
+import { BegrunnelsesFelt } from '../Felles/BegrunnelsesFelt';
+import { AlertError } from '../../../../../Felles/Visningskomponenter/Alerts';
+import HovedKnapp from '../../../../../Felles/Knapper/HovedKnapp';
+import { validerSkoleårsperioderMedBegrunnelse } from '../Felles/vedtaksvalidering';
 
-const Skoleårsperiode = styled.div`
+const Form = styled.form`
     display: flex;
     flex-direction: column;
-    row-gap: 0.75rem;
-    padding: 1rem;
-    background-color: ${AGray50};
+    gap: 1rem;
 `;
 
-const TilbakestillButton = styled(TilbakestillKnapp)`
-    margin-bottom: 0.25rem;
+const Utregningstabell = styled(UtregningstabellSkolepenger)`
+    margin-left: 1rem;
 `;
 
-interface Props {
-    skoleårsperioder: ListState<ISkoleårsperiodeSkolepenger>;
-    skoleårsperioderForrigeVedtak: ISkoleårsperiodeSkolepenger[];
-    valideringsfeil: FormErrors<InnvilgeVedtakForm>['skoleårsperioder'];
-    settValideringsFeil: Dispatch<SetStateAction<FormErrors<InnvilgeVedtakForm>>>;
-}
-
-const beregnSkoleårForSkoleårsperiode = (periode: ISkoleårsperiodeSkolepenger) => {
-    return (
-        beregnSkoleår(
-            periode.perioder[0].årMånedFra,
-            periode.perioder[0].årMånedTil
-        ) as GyldigBeregnetSkoleår
-    ).skoleår;
+export const defaultSkoleårsperioder = (
+    forrigeVedtak?: IvedtakForSkolepenger
+): ISkoleårsperiodeSkolepenger[] => {
+    const forrigeSkoleårsperioder = forrigeVedtak?.skoleårsperioder;
+    if (forrigeSkoleårsperioder && forrigeSkoleårsperioder.length > 0) {
+        return forrigeSkoleårsperioder;
+    } else {
+        return [tomSkoleårsperiodeSkolepenger()];
+    }
 };
 
-const OpphøreVedtak: React.FC<Props> = ({
-    skoleårsperioder,
-    skoleårsperioderForrigeVedtak,
-    valideringsfeil,
-    settValideringsFeil,
-}) => {
-    const { behandlingErRedigerbar } = useBehandling();
-    const { settIkkePersistertKomponent } = useApp();
+export const OpphøreVedtak: React.FC<{
+    behandling: Behandling;
+    lagretInnvilgetVedtak?: IvedtakForSkolepenger;
+    forrigeVedtak?: IvedtakForSkolepenger;
+}> = ({ behandling, lagretInnvilgetVedtak, forrigeVedtak }) => {
+    const { behandlingErRedigerbar, hentBehandling } = useBehandling();
+    const { axiosRequest, nullstillIkkePersisterteKomponenter, settIkkePersistertKomponent } =
+        useApp();
+    const { utførRedirect } = useRedirectEtterLagring(`/behandling/${behandling.id}/simulering`);
 
-    const fjernSkoleårsperiode = (skoleårsperiode: ISkoleårsperiodeSkolepenger) => {
-        const index = skoleårsperioder.value.indexOf(skoleårsperiode);
-        skoleårsperioder.remove(index);
-        settIkkePersistertKomponent(VEDTAK_OG_BEREGNING);
-    };
-
-    const tilbakestillSkoleårsperiode = (forrigeIndex: number) => {
-        const indexForElementFørId = locateIndexToRestorePreviousItemInCurrentItems(
-            forrigeIndex,
-            skoleårsperioder.value,
-            skoleårsperioderForrigeVedtak,
-            (t1, t2) => beregnSkoleårForSkoleårsperiode(t1) === beregnSkoleårForSkoleårsperiode(t2)
-        );
-        skoleårsperioder.setValue((prevState) => {
-            return [
-                ...prevState.slice(0, indexForElementFørId),
-                skoleårsperioderForrigeVedtak[forrigeIndex],
-                ...prevState.slice(indexForElementFørId, prevState.length),
-            ];
-        });
-        settIkkePersistertKomponent(VEDTAK_OG_BEREGNING);
-    };
-
-    /**
-     * Sender med previousValue som er item i data som skal oppdateres
-     * Denne brukes for å finne index i skoleårsperioder
-     */
-    const oppdaterSkoleårsperioder = <T extends ISkoleårsperiodeSkolepenger>(
-        previousValue: T,
-        property: keyof T,
-        value: T[keyof T]
-    ) => {
-        const index = skoleårsperioder.value.indexOf(previousValue);
-        const skoleårsperiode = skoleårsperioder.value[index];
-        skoleårsperioder.update({ ...skoleårsperiode, [property]: value }, index);
-        settIkkePersistertKomponent(VEDTAK_OG_BEREGNING);
-    };
-
-    /**
-     * For å sette valideringsfeil på riktig indeks returneres periode og index til skoleårsperioden for nye perioder
-     */
-    const skoleårsperioderPerSkoleår = useMemo(
-        () =>
-            skoleårsperioder.value.reduce(
-                (acc, periode, index) => {
-                    if (periode.perioder.length > 0) {
-                        return {
-                            ...acc,
-                            [beregnSkoleårForSkoleårsperiode(periode)]: { periode, index },
-                        };
-                    } else {
-                        return acc;
-                    }
-                },
-                {} as Record<
-                    number,
-                    { periode: ISkoleårsperiodeSkolepenger; index: number } | undefined
-                >
-            ),
-        [skoleårsperioder]
+    const [laster, settLaster] = useState<boolean>(false);
+    const [feilmelding, settFeilmelding] = useState('');
+    const [beregningsresultat, settBeregningsresultat] = useState(
+        byggTomRessurs<IBeregningSkolepengerResponse>()
     );
+
+    const formState = useFormState<InnvilgeVedtakForm>(
+        {
+            skoleårsperioder: lagretInnvilgetVedtak
+                ? lagretInnvilgetVedtak.skoleårsperioder
+                : defaultSkoleårsperioder(forrigeVedtak),
+            begrunnelse: lagretInnvilgetVedtak?.begrunnelse || '',
+        },
+        validerSkoleårsperioderMedBegrunnelse
+    );
+
+    const skoleårsPerioderState = formState.getProps(
+        'skoleårsperioder'
+    ) as ListState<ISkoleårsperiodeSkolepenger>;
+
+    const begrunnelseState = formState.getProps('begrunnelse') as FieldState;
+
+    const lagreVedtak = (vedtaksRequest: IVedtakForSkolepenger) => {
+        settLaster(true);
+        nullstillIkkePersisterteKomponenter();
+        axiosRequest<string, IVedtakForSkolepenger>({
+            method: 'POST',
+            url: `/familie-ef-sak/api/vedtak/${behandling.id}/lagre-vedtak`,
+            data: vedtaksRequest,
+        })
+            .then(håndterVedtaksresultat())
+            .finally(() => {
+                settLaster(false);
+            });
+    };
+
+    const håndterVedtaksresultat = () => {
+        return (res: Ressurs<string>) => {
+            switch (res.status) {
+                case RessursStatus.SUKSESS:
+                    utførRedirect();
+                    hentBehandling.rerun();
+                    break;
+                case RessursStatus.HENTER:
+                case RessursStatus.IKKE_HENTET:
+                    break;
+                default:
+                    settIkkePersistertKomponent(uuidv4());
+                    settFeilmelding(res.frontendFeilmelding);
+            }
+        };
+    };
+
+    const handleSubmit = (form: FormState<InnvilgeVedtakForm>) => {
+        const vedtaksRequest: IVedtakForSkolepenger = {
+            skoleårsperioder: form.skoleårsperioder,
+            begrunnelse: form.begrunnelse,
+            _type: IVedtakType.OpphørSkolepenger,
+            resultatType: EBehandlingResultat.OPPHØRT,
+        };
+        lagreVedtak(vedtaksRequest);
+    };
+
+    useEffect(() => {
+        if (!behandlingErRedigerbar) {
+            axiosRequest<IBeregningSkolepengerResponse, null>({
+                method: 'GET',
+                url: `/familie-ef-sak/api/beregning/skolepenger/${behandling.id}`,
+            }).then((res: Ressurs<IBeregningSkolepengerResponse>) => settBeregningsresultat(res));
+        }
+    }, [axiosRequest, behandling, behandlingErRedigerbar]);
 
     return (
-        <>
-            {skoleårsperioderForrigeVedtak.map((forrigeSkoleårsperiode, index) => {
-                const skoleår = beregnSkoleårForSkoleårsperiode(forrigeSkoleårsperiode);
-                const skoleårsperiode = skoleårsperioderPerSkoleår[skoleår];
-                const skoleårsperiodeErOpphørt = !skoleårsperiode;
-                return (
-                    <Skoleårsperiode key={index}>
-                        <Delårsperioder
-                            data={
-                                skoleårsperiodeErOpphørt
-                                    ? forrigeSkoleårsperiode.perioder
-                                    : skoleårsperiode.periode.perioder
-                            }
-                            oppdater={(perioder) =>
-                                skoleårsperiode &&
-                                oppdaterSkoleårsperioder(
-                                    skoleårsperiode.periode,
-                                    'perioder',
-                                    perioder
-                                )
-                            }
-                            behandlingErRedigerbar={
-                                behandlingErRedigerbar && !skoleårsperiodeErOpphørt
-                            }
-                            valideringsfeil={
-                                valideringsfeil &&
-                                skoleårsperiode &&
-                                valideringsfeil[skoleårsperiode.index]?.perioder
-                            }
-                            settValideringsFeil={(oppdaterteFeil) =>
-                                skoleårsperiode &&
-                                oppdaterValideringsfeil(
-                                    settValideringsFeil,
-                                    skoleårsperiode.index,
-                                    'perioder',
-                                    oppdaterteFeil
-                                )
-                            }
-                            erSkoleårOpphørt={skoleårsperiodeErOpphørt}
-                            erOpphør={true}
-                        />
-                        <Utgiftsperioder
-                            behandlingErRedigerbar={behandlingErRedigerbar}
-                            data={skoleårsperiode?.periode?.utgiftsperioder || []}
-                            forrigeData={forrigeSkoleårsperiode.utgiftsperioder}
-                            oppdater={(utgiftsperioder) =>
-                                skoleårsperiode &&
-                                oppdaterSkoleårsperioder(
-                                    skoleårsperiode.periode,
-                                    'utgiftsperioder',
-                                    utgiftsperioder
-                                )
-                            }
-                            skoleårErOpphørt={skoleårsperiodeErOpphørt}
-                        />
-                        {behandlingErRedigerbar && !skoleårsperiodeErOpphørt && (
-                            <FjernKnapp
-                                onClick={() => fjernSkoleårsperiode(skoleårsperiode.periode)}
-                                knappetekst={'Fjern skoleårsperide'}
-                            />
-                        )}
-                        {behandlingErRedigerbar && skoleårsperiodeErOpphørt && (
-                            <TilbakestillButton
-                                onClick={() => tilbakestillSkoleårsperiode(index)}
-                                knappetekst={'Tilbakestill skoleår'}
-                            />
-                        )}
-                    </Skoleårsperiode>
-                );
-            })}
-        </>
+        <Form onSubmit={formState.onSubmit(handleSubmit)}>
+            <BegrunnelsesFelt begrunnelseState={begrunnelseState} errorState={formState.errors} />
+            <OpphøreSkolepenger
+                skoleårsperioder={skoleårsPerioderState}
+                skoleårsperioderForrigeVedtak={forrigeVedtak?.skoleårsperioder || []}
+                valideringsfeil={formState.errors.skoleårsperioder}
+                settValideringsFeil={formState.setErrors}
+            />
+            {feilmelding && <AlertError>{feilmelding}</AlertError>}
+            <Utregningstabell beregningsresultat={beregningsresultat} />
+            {behandlingErRedigerbar && <HovedKnapp disabled={laster} knappetekst="Lagre vedtak" />}
+        </Form>
     );
 };
-
-export default OpphøreVedtak;
